requirements.txt
Flask
pymongo
python-dotenv
requests
pandas
numpy
joblib
scikit-learn
statsmodels
xgboost
